requirements.txt
openai
python-dotenv
lancedb
PyMuPDF
boto3
botocore
pandas
pyarrow
httpx